example: clear BT RX buffer after processing
the way the buffer is used in the DMA operation means that it's never
cleared out. as long as the new messages are the same length (or longer)
than the previous ones means that they just overwrite each other and
everything is fine.

but if the previous message was longer (which can happen if multiple
button presses happen fast enough one after the other, thus leading to
no line idle interrupt in-between) the next message will only overwrite
the first command(s) and the other old commands will stay in there and
be processed again.